(task) [WIP] Create modal for entering data sources
