+++ webapp/src/components/Grid/TableHead.js
@@ -13,26 +13,16 @@ export default () => {
   const schema = useSchema((state) => state[querySpecification.sourceLabel]);
   const tableColors = querySpecification.tableColors;
   const headList = [];
-  let selectedTableNames = [],
-    selectedTableColumNames = [],
+  let selectedTableColumNames = [],
     _columns = [];
+
   if ("isEmbedded" in queryContext && queryContext.isEmbedded) {
-    selectedTableNames = [
-      ...new Set(
-        querySpecification.embeddedColumns[queryContext.embeddedDataIndex].map(
-          (x) => x.tableName
-        )
-      ),
-    ];
     selectedTableColumNames = querySpecification.embeddedColumns[
       queryContext.embeddedDataIndex
     ].map((x) => x.label);
     _columns =
       querySpecification.embeddedColumns[queryContext.embeddedDataIndex];
   } else {
-    selectedTableNames = [
-      ...new Set(querySpecification.columns.map((x) => x.tableName)),
-    ];
     selectedTableColumNames = querySpecification.columns.map((x) => x.label);
     _columns = querySpecification.columns;
   }
