+++ webapp/src/services/global/constants.js
@@ -4,3 +4,8 @@ export const APP_NAME_BROWSER = "APP_NAME_BROWSER";
 export const APP_NAME_FUNNEL = "APP_NAME_FUNNEL";
 export const APP_NAME_REPORT = "APP_NAME_REPORT";
 export const APP_NAME_CALENDAR = "APP_NAME_CALENDAR";
+
+export const OBJECT_CREATE = "OBJECT_CREATE";
+export const OBJECT_READ = "OBJECT_READ";
+export const OBJECT_UPDATE = "OBJECT_UPDATE";
+export const OBJECT_DELETE = "OBJECT_DELETE";
